+++ plugin/src/main/java/com/stratio/cassandra/lucene/search/condition/BitemporalCondition.java
@@ -27,6 +27,7 @@ import static com.stratio.cassandra.lucene.schema.mapping.BitemporalMapper.Bitem
 import static org.apache.lucene.search.BooleanClause.Occur.MUST;
 import static org.apache.lucene.search.BooleanClause.Occur.SHOULD;
 import static org.apache.lucene.search.NumericRangeQuery.newLongRange;
+import static com.stratio.cassandra.lucene.schema.mapping.BitemporalMapper.*;
 
 /**
  * A {@link Condition} implementation that matches bi-temporal (four) fields within two range of values.